frontend/Vishal-M/speech_transcription_platform/backend/requirements.txt
# Python Version
# Requires Python 3.10
python-version >= "3.10, <3.11"

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Azure Services
azure-cognitiveservices-speech==1.34.0

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Async & WebSocket
websockets==12.0
aiofiles==23.2.1

# Monitoring & Logging
prometheus-client==0.19.0
structlog==24.1.0

# Security
python-multipart==0.0.6
slowapi==0.1.9

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
black==24.1.0
ruff==0.1.13
